Done with Cucumber

+++ db/migrate/20121208201741_create_foods.rb
@@ -2,7 +2,7 @@ class CreateFoods < ActiveRecord::Migration
   def change
     create_table :foods do |t|
       t.string :name
-      t.float :price
+      t.decimal :price
       t.boolean :available
 
       t.timestamps
